docGen/docgen/requirements.txt
openai>=0.27.0
gitpython>=3.1.30